xr_classifier/requirements.txt
fastapi>=0.111,<0.133
uvicorn[standard]>=0.30
pillow>=10.0,<11.0
numpy>=1.26,<2.0
scikit-learn>=1.4,<2.0
albumentation>=0.14
opencv-python-headless>=4.9
torch>=2.0,<3.0
torchvision>=0.15,<1.0
PyYAML>=0.46,<1.0
# core
fastai==2.7.12
timm==0.9.2
pytorch-grad-cam==1.4.5
scikit-learn
pandas
matplotlib
fastapi
uvicorn[standard]
pydantic
torch
torchvision
fastai
numpy
pandas
matplotlib
seaborn
scikit-learn
torchmetrics

seaborn
# PyTorch: choose one of the following lines depending on your machine
# CPU-only PyTorch (uncomment the CPU line if you don't have a CUDA GPU)
# torch==2.2.0+cpu
# torchvision==0.16.0+cpu
# torchaudio==2.2.0+cpu
# pip install command for CPU:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

# CUDA 11.8 (common), uncomment & run this instead of above CPU lines:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# other packages installed from PyPI
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distlib==0.4.0
distro==1.9.0
executing==2.2.0
filelock==3.19.1
fonttools==4.58.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.13
idna==3.10
imbalanced-learn==0.13.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
narwhals==1.44.0
nest-asyncio==1.6.0
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.6
openai==1.88.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.0
parso==0.8.4
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
plotly-express==0.4.1
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
PyAudio==0.2.14
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
statsmodels==0.14.4
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
wcwidth==0.2.13
wikipedia==1.4.0
wordcloud==1.9.4
xgboost==3.0.2
